extract MarkdownView
