+++ samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/viewmodel/MainViewModel.kt
@@ -5,10 +5,20 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.graphics.Color
-import org.jetbrains.jewel.samples.standalone.IntUiThemes
+import org.jetbrains.jewel.intui.markdown.standalone.dark
+import org.jetbrains.jewel.intui.markdown.standalone.light
+import org.jetbrains.jewel.intui.markdown.standalone.styling.dark
+import org.jetbrains.jewel.intui.markdown.standalone.styling.extensions.github.alerts.dark
+import org.jetbrains.jewel.intui.markdown.standalone.styling.extensions.github.alerts.light
+import org.jetbrains.jewel.intui.markdown.standalone.styling.light
+import org.jetbrains.jewel.markdown.extensions.github.alerts.AlertStyling
+import org.jetbrains.jewel.markdown.extensions.github.alerts.GitHubAlertRendererExtension
+import org.jetbrains.jewel.markdown.rendering.MarkdownBlockRenderer
+import org.jetbrains.jewel.markdown.rendering.MarkdownStyling
 import org.jetbrains.jewel.samples.showcase.components.StandaloneSampleIcons
+import org.jetbrains.jewel.samples.showcase.views.MarkdownDemo
+import org.jetbrains.jewel.samples.standalone.IntUiThemes
 import org.jetbrains.jewel.samples.standalone.view.ComponentsView
-import org.jetbrains.jewel.samples.standalone.view.MarkdownDemo
 import org.jetbrains.jewel.samples.standalone.view.WelcomeView
 
 object MainViewModel {
@@ -52,6 +62,23 @@ private val mainMenuItems =
             iconKey = StandaloneSampleIcons.markdown,
             keyboardShortcut =
                 KeyBinding(macOs = setOf("⌥", "M"), windows = setOf("Alt", "M"), linux = setOf("Alt", "M")),
-            content = { MarkdownDemo() },
+            content = {
+                MarkdownDemo(
+                    darkStyling = MarkdownStyling.dark(),
+                    darkRenderer =
+                        MarkdownBlockRenderer.dark(
+                            styling = MarkdownStyling.dark(),
+                            rendererExtensions =
+                                listOf(GitHubAlertRendererExtension(AlertStyling.dark(), MarkdownStyling.dark())),
+                        ),
+                    lightStyling = MarkdownStyling.light(),
+                    lightRenderer =
+                        MarkdownBlockRenderer.light(
+                            styling = MarkdownStyling.light(),
+                            rendererExtensions =
+                                listOf(GitHubAlertRendererExtension(AlertStyling.light(), MarkdownStyling.light())),
+                        ),
+                )
+            },
         ),
     )
